+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -134,7 +134,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-12T12:44:33.556Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-12T13:06:03.318Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")