+++ packages/orq-rc/RELEASES.md
@@ -809,3 +809,13 @@ Based on:
 - [python v3.2.0-rc.43] packages/orq-rc
 ### Releases
 - [PyPI v3.2.0-rc.43] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.43 - packages/orq-rc
+
+## 2025-03-12 13:07:16
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.517.0 (2.548.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.2.0-rc.44] packages/orq-rc
+### Releases
+- [PyPI v3.2.0-rc.44] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.44 - packages/orq-rc